backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.1
billiard==4.2.4
black==26.3.1
celery==5.6.3
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==6.0.3
django-cors-headers==4.9.0
django-redis==6.0.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
iniconfig==2.3.0
kombu==5.6.2
mypy_extensions==1.1.0
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytokens==0.4.1
redis==7.4.0
six==1.17.0
sqlparse==0.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
